остались role and user

+++ Api budger/Services/BudgerService.cs
@@ -48,7 +48,7 @@ namespace Api_budger.Services
 
             if (!listUser.Any(u => u == userId))
             {
-                throw new UnauthorizedAccessException("У вас нет прав для удаления категорий доходов этой семьи.");
+                throw new UnauthorizedAccessException("У вас нет прав для удаления расходов этой семьи.");
             }
 
 
@@ -75,21 +75,47 @@ namespace Api_budger.Services
             return result;
         }
 
-        public Task<IEnumerable<Incom>> AddIncomsAsyns(IEnumerable<InputIncom> inputIncom)
+        public async Task<IEnumerable<Incom>> AddIncomsAsyns(IEnumerable<InputIncom> inputIncom)
         {
-            var incom = _mapper.Map<IEnumerable<Incom>>(inputIncom);
+            var userId = _currentUserService.GetUserId();
+            if (userId == 0) throw new KeyNotFoundException("Пользователь не найден");
+
+            List<long> listUser = new List<long>();
+
+            //TODO: рефактор логики (возможно когда 1 из списка подходит то if пропустит все остальные)
+            foreach (var incomCategory in inputIncom)
+            {
+                listUser.AddRange(await _budgerRepository.GetUserIdByIncomCategoryId(incomCategory.IncomeCategoryId));
+            }
+            listUser = (List<long>)listUser.Distinct();
+
+
+            if (!listUser.Any(u => u == userId))
+            {
+                throw new UnauthorizedAccessException("У вас нет прав для удаления доходов этой семьи.");
+            }
+
+
+            var incoms = _mapper.Map<IEnumerable<Incom>>(inputIncom);
             var date = DateTime.UtcNow;
-            incom.ToList().ForEach(b => b.Date = date);
-            var result = _budgerRepository.AddIncomsAsyns(incom);
+
+            incoms.ToList().ForEach(b => b.Date = date);
+            incoms.ToList().ForEach(u => u.UserId = userId);
+
+            var result = await _budgerRepository.AddIncomsAsyns(incoms);
             return result;
         }
 
         public async Task<IEnumerable<IncomCategory>> AddIncomCategoryInFamilyAsyns(IEnumerable<InputIncomCategory> inputIncomCategory)
         {
-            long familiId = -1;
-            if (familiId <= 0) throw new Exception("famili Id <= 0");
+            var userId = _currentUserService.GetUserId();
+            if (userId == 0) throw new KeyNotFoundException("Пользователь не найден");
+
+            var familyId = await _userRepository.GetFamilyIdByUserIdAsync(userId);
+            if (familyId <= 0) throw new KeyNotFoundException("Семья не найдена");
+
             var incomCategoryList = _mapper.Map<IEnumerable<IncomCategory>>(inputIncomCategory);
-            var result = await _budgerRepository.AddIncomCategoryInFamilyAsyns(familiId, incomCategoryList);
+            var result = await _budgerRepository.AddIncomCategoryInFamilyAsyns(familyId, incomCategoryList);
             return result;
         }
 
@@ -102,7 +128,7 @@ namespace Api_budger.Services
 
             if (!userIdByBC.Contains(userId))
             {
-                throw new UnauthorizedAccessException("У вас нет прав для удаления категорий доходов этой семьи.");
+                throw new UnauthorizedAccessException("У вас нет прав для удаления расходов этой семьи.");
             }
 
             var newBudger = _mapper.Map<Budger>(inputBudger);
@@ -132,17 +158,38 @@ namespace Api_budger.Services
 
         public async Task<Incom> CorrectIncomAsyns(long id, InputIncom inputIncom)
         {
-            var incom = _mapper.Map<Incom>(inputIncom);
-            var result = await _budgerRepository.CorrectIncomAsyns(id, incom);
+            var userId = _currentUserService.GetUserId();
+            if (userId == 0) throw new KeyNotFoundException("Пользователь не найден");
+
+            var userIdByIC = await _budgerRepository.GetUserIdByIncomCategoryId(id);
+
+            if (!userIdByIC.Contains(userId))
+            {
+                throw new UnauthorizedAccessException("У вас нет прав для удаления категорий доходов этой семьи.");
+            }
+
+            var newIncom = _mapper.Map<Incom>(inputIncom);
+            var result = await _budgerRepository.CorrectIncomAsyns(id, newIncom);
             if (result is null) throw new Exception("New incom incorrect");
             return result;
         }
 
-        public async Task<IncomCategory> CorrectIncomCategoryFromUserByIdAsyns(long id, long userId, InputIncomCategory inputIncomCategory)
+        public async Task<IncomCategory> CorrectIncomCategoryFromUserByIdAsyns(long id, InputIncomCategory inputIncomCategory)
         {
+            var userId = _currentUserService.GetUserId();
+            if (userId == 0) throw new KeyNotFoundException("Пользователь не найден");
+
+            var userIdByBudgerCategoryId = await _budgerRepository.GetUserIdByIncomCategoryId(id);
+
+            if (!userIdByBudgerCategoryId.Any(u => u == userId))
+            {
+                throw new UnauthorizedAccessException("У вас нет прав для удаления категорий доходов этой семьи.");
+            }
+
             var newIncomCategory = _mapper.Map<IncomCategory>(inputIncomCategory);
+
             var result = await _budgerRepository.CorrectIncomCategoryFromUserByIdAsyns(id, userId, newIncomCategory);
-            if (result is null) throw new Exception("New incom category incorrect");
+            if (result is null) throw new Exception("Новая категория доходов не корректна или не верна");
             return result;
         }
 
@@ -185,11 +232,38 @@ namespace Api_budger.Services
 
         public async Task<bool> DeleteIncomByIdAsyns(long incomId)
         {
+            var currentUserId = _currentUserService.GetUserId();
+            if (currentUserId == 0) throw new KeyNotFoundException("Пользователь не найден");
+
+            var familyId = await _userRepository.GetFamilyIdByUserIdAsync(currentUserId);
+            if (familyId <= 0) throw new KeyNotFoundException("Семья не найдена");
+
+            var userIdByBudgerCategoryId = await _budgerRepository.GetUserIdsByIncomIdAsyns(incomId);
+            if (userIdByBudgerCategoryId.IsNullOrEmpty()) throw new KeyNotFoundException("Пользователь не найден");
+
+            if (!(userIdByBudgerCategoryId.Contains(currentUserId)))
+            {
+                throw new UnauthorizedAccessException("У вас нет прав для удаления доходов этой семьи.");
+            }
+
             return await _budgerRepository.DeleteIncomByIdAsyns(incomId);
         }
 
-        public async Task<bool> DeleteIncomCategoryFromFamilyByIdAsyns(long incomCategoryId, long familyId)
+        public async Task<bool> DeleteIncomCategoryFromFamilyByIdAsyns(long incomCategoryId)
         {
+            var userId = _currentUserService.GetUserId();
+            if (userId == 0) throw new KeyNotFoundException("Пользователь не найден");
+
+            var familyId = await _userRepository.GetFamilyIdByUserIdAsync(userId);
+            if (familyId <= 0) throw new KeyNotFoundException("Семья не найдена");
+
+            var userIdByBudgerCategoryId = await _budgerRepository.GetUserIdByIncomCategoryId(incomCategoryId);
+
+            if (!userIdByBudgerCategoryId.Any(u => u == userId))
+            {
+                throw new UnauthorizedAccessException("У вас нет прав для удаления категорий доходов этой семьи.");
+            }
+
             return await _budgerRepository.DeleteIncomCategoryFromFamilyByIdAsyns(incomCategoryId, familyId);
         }
 
@@ -228,50 +302,35 @@ namespace Api_budger.Services
             return result;
         }
 
-        public async Task<IEnumerable<Incom>> GetIncomByFamilyIdAsyns(long familyId)
+        public async Task<IEnumerable<Incom>> GetIncomByFamilyIdAsyns()
         {
-            var userIdByFamily = await _userRepository.GetUserIdByFamilyAsync(familyId);
             var userId = _currentUserService.GetUserId();
-            if (userIdByFamily.IsNullOrEmpty()) throw new KeyNotFoundException("User not found");
-            var userRole = _currentUserService.GetUserRole();
+            if (userId == 0) throw new KeyNotFoundException("Пользователь не найден");
 
-            if (!(userIdByFamily.Any(u => u == userId) || userRole == Consts.RoleConst.admin))
-            {
-                throw new UnauthorizedAccessException("У вас нет прав для просмотра доходов этой семьи.");
-            }
+            var familyId = await _userRepository.GetFamilyIdByUserIdAsync(userId);
+            if (familyId <= 0) throw new KeyNotFoundException("Семья не найден");
 
             var result = await _budgerRepository.GetIncomByFamilyIdAsyns(familyId);
             if (result is null) throw new Exception("Доходы пусты");
             return result;
         }
 
-        public async Task<IEnumerable<Incom>> GetIncomByUserIdAsyns(long userId)
+        public async Task<IEnumerable<Incom>> GetIncomByUserIdAsyns()
         {
             var cyrrentUserId = _currentUserService.GetUserId();
-            if (cyrrentUserId == 0) throw new KeyNotFoundException("User not found");
-            var userRole = _currentUserService.GetUserRole();
+            if (cyrrentUserId <= 0) throw new KeyNotFoundException("Пользователь не найден");
 
-            if (!(userId == cyrrentUserId || userRole == Consts.RoleConst.admin))
-            {
-                throw new UnauthorizedAccessException("У вас нет прав для просмотра доходов этого пользователя");
-            }
-
-            var result = await _budgerRepository.GetIncomByUserIdAsyns(userId);
+            var result = await _budgerRepository.GetIncomByUserIdAsyns(cyrrentUserId);
             if (result is null) throw new Exception("Доходы пусты");
             return result;
         }
 
-        public async Task<IEnumerable<IncomCategory>> GetIncomCategoryByFamilyIdAsyns(long familyId)
+        public async Task<IEnumerable<IncomCategory>> GetIncomCategoryByFamilyIdAsyns()
         {
-            var userIdByFamily = await _userRepository.GetUserIdByFamilyAsync(familyId);
             var userId = _currentUserService.GetUserId();
-            if (userIdByFamily.IsNullOrEmpty()) throw new KeyNotFoundException("User not found");
-            var userRole = _currentUserService.GetUserRole();
+            if (userId <= 0) throw new KeyNotFoundException("Пользователь не найден");
 
-            if (!(userIdByFamily.Any(u => u == userId) || userRole == Consts.RoleConst.admin))
-            {
-                throw new UnauthorizedAccessException("У вас нет прав для просмотра категории доходов этой семьи.");
-            }
+            var familyId = await _userRepository.GetFamilyIdByUserIdAsync(userId);
 
             var result = await _budgerRepository.GetIncomCategoryByFamilyIdAsyns(familyId);
             if (result is null) throw new Exception("Категория доходов пуста");
